Use Cells endpoint from bbsClient
[#103637092]

+++ maintain/maintain.go
@@ -100,7 +100,7 @@ func (m *Maintainer) createHeartbeater() (ifrit.Runner, error) {
 		return nil, err
 	}
 
-	cellCapacity := models.NewCellCapacity(resources.MemoryMB, resources.DiskMB, resources.Containers)
+	cellCapacity := models.NewCellCapacity(int32(resources.MemoryMB), int32(resources.DiskMB), int32(resources.Containers))
 	cellPresence := models.NewCellPresence(m.CellID, m.RepAddress, m.Zone, cellCapacity, m.RootFSProviders, m.PreloadedRootFSes)
 	return m.serviceClient.NewCellPresenceRunner(m.logger, &cellPresence, m.RetryInterval), nil
 }
